requirement.txt
# 核心框架
flask>=2.0.0
flask-cors>=4.0.0

# 数据处理
numpy>=1.21.0
pillow>=8.0.0
tqdm>=4.62.0

# 深度学习
torch>=1.10.0
torchvision>=0.11.0

# 向量检索
faiss-cpu>=1.7.0

# 任务队列
redis>=4.0.0
celery>=5.2.0

# 数据库
sqlite3  # Python内置

# 开发工具
pytest>=6.0.0